move storage account service to service package
