+++ src/main/java/it/gov/pagopa/rtd/ms/rtdmsfilereporter/domain/service/StorageAccountService.java
@@ -1,6 +1,8 @@
-package it.gov.pagopa.rtd.ms.rtdmsfilereporter.feign;
+package it.gov.pagopa.rtd.ms.rtdmsfilereporter.domain.service;
 
 import it.gov.pagopa.rtd.ms.rtdmsfilereporter.domain.model.AggregatesDataSummary;
+import it.gov.pagopa.rtd.ms.rtdmsfilereporter.feign.AggregatesDataSummaryMapper;
+import it.gov.pagopa.rtd.ms.rtdmsfilereporter.feign.StorageAccountRestConnector;
 import java.io.IOException;
 import java.util.Map;
 import lombok.RequiredArgsConstructor;
